+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Upcoming release
+
+- Preload in-app message's images
+
 ## 3.9.0
 
 - Add support for deferred links